requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
Django==3.0.5
django-allauth==0.41.0
django-cors-headers==3.2.1
django-filter==2.2.0
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
djangorestframework==3.11.0
idna==2.9
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
oauthlib==3.1.0
openapi-codec==1.3.2
Pillow==7.1.1
pybase64==1.0.1
PyJWT==1.6.4
pyotp==2.3.0
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
simplejson==3.17.0
six==1.14.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.9
whitenoise